chore: Added parsed redeemer info to transaction_scripts

+++ stores/script/src/main/java/com/bloxbean/cardano/yaci/store/script/domain/Redeemer.java
@@ -10,8 +10,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
@@ -20,7 +18,7 @@ import java.math.BigInteger;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class Redeemer {
     private RedeemerTag tag;
-    private BigInteger index;
+    private Integer index;
     private String data;
     private ExUnits exUnits;
 }
